python_ext/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2
gunicorn==21.2.0 